+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 **Features:**
 
 - Add a config option to add default tags to all Relay Sentry events. ([#3944](https://github.com/getsentry/relay/pull/3944))
+- Automatically derive `client.address` and `user.geo` for standalone spans. ([#4047](https://github.com/getsentry/relay/pull/4047))
 
 ## 24.9.0